+++ lib/models/cell_data.dart
@@ -18,7 +18,6 @@ extension AreaDensityExtension on AreaDensity {
 
 class CellData {
   final int? sectionNumber;
-  final LocationType locationType;
   final bool pirates;
   final String megacorporation;
 
@@ -26,7 +25,6 @@ class CellData {
 
   const CellData({
     this.sectionNumber,
-    this.locationType = LocationType.world,
     this.pirates = false,
     this.megacorporation = '',
   });
@@ -34,19 +32,16 @@ class CellData {
   const CellData.deepSpace({
     this.pirates = false,
     this.megacorporation = '',
-  })  : sectionNumber = null,
-        locationType = LocationType.world;
+  }) : sectionNumber = null;
 
   CellData copyWith({
     int? sectionNumber,
-    LocationType? locationType,
     bool? pirates,
     String? megacorporation,
     bool clearSection = false,
   }) {
     return CellData(
       sectionNumber: clearSection ? null : (sectionNumber ?? this.sectionNumber),
-      locationType: locationType ?? this.locationType,
       pirates: pirates ?? this.pirates,
       megacorporation: megacorporation ?? this.megacorporation,
     );
@@ -54,14 +49,12 @@ class CellData {
 
   Map<String, dynamic> toJson() => {
     'sectionNumber': sectionNumber,
-    'locationType': locationType.toJson(),
     'pirates': pirates,
     'megacorporation': megacorporation,
   };
 
   factory CellData.fromJson(Map<String, dynamic> json) => CellData(
     sectionNumber: json['sectionNumber'] as int?,
-    locationType: LocationTypeExtension.fromJson(json['locationType'] as String),
     pirates: json['pirates'] as bool? ?? false,
     megacorporation: json['megacorporation'] as String? ?? '',
   );